KienPT commit update source entity

+++ src/main/java/com/fptu/edu/travelservices/entity/User.java
@@ -22,7 +22,7 @@ import java.util.Set;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Integer id;
 
     @Column(name = "FirstName")
     private String firstName;
